Fix create_table migration rule

+++ entity/src/customization.rs
@@ -2,14 +2,15 @@
 
 use sea_orm::entity::prelude::*;
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "customization")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub bowl_id: i32,
     pub name: String,
-    pub price: Decimal,
+    #[sea_orm(column_type = "Float")]
+    pub price: f32,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
